* MISUV-9397: ** - drop validChargeTypeCondition method from the main scope;
