+++ it/test/controllers/ControllerISpecHelper.scala
@@ -23,9 +23,9 @@ import helpers.ComponentSpecBase
 import helpers.servicemocks.BusinessDetailsStub.stubGetBusinessDetails
 import helpers.servicemocks.CitizenDetailsStub.stubGetCitizenDetails
 import helpers.servicemocks._
-import models.FinancialDetailsModelExtension
 import models.admin.FeatureSwitchName
 import models.admin.FeatureSwitchName.allFeatureSwitches
+import models.extensions.FinancialDetailsModelExtension
 import play.api.http.Status.{SEE_OTHER, UNAUTHORIZED}
 import testConstants.BaseIntegrationTestConstants.getAgentClientDetailsForCookie
 import testOnly.repository.FeatureSwitchRepository
