+++ test/models/extensions/FinancialDetailsModelExtension.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package models
+package models.extensions
 
+import enums.{Poa1Charge, Poa1ReconciliationDebit, Poa2Charge, Poa2ReconciliationDebit, TRMAmendCharge, TRMNewCharge}
 import models.financialDetails.{DocumentDetail, FinancialDetailsModel}
 
 import java.time.LocalDate
@@ -34,6 +35,14 @@ trait FinancialDetailsModelExtension {
       fdm match {
         case FinancialDetailsModel(_, documentDetails, _) =>
           documentDetails.flatMap(_.getDueDate())
+      }
+    }
+
+    def validChargeTypeCondition: DocumentDetail => Boolean = documentDetail => {
+      (documentDetail.documentText, documentDetail.getDocType) match {
+        case (Some(documentText), _) if documentText.contains("Class 2 National Insurance") => true
+        case (_, Poa1Charge | Poa2Charge | Poa1ReconciliationDebit | Poa2ReconciliationDebit | TRMNewCharge | TRMAmendCharge) => true
+        case (_, _) => false
       }
     }
   }
